requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
anthropic>=0.49.0
docling>=2.15.0
python-dotenv>=1.0.0
httpx>=0.27.0
